cleanup the login part of e2e
