+++ packages/e2e/e2e-base/package.json
@@ -36,6 +36,7 @@
     "@rushstack/eslint-config": "~3.6.9",
     "@types/node": "^20.12.7",
     "directory-tree": "3.5.1",
+    "dotenv": "~16.4.5",
     "eslint": "^8.45.0",
     "eslint-plugin-playwright": "1.6.2",
     "prettier": "~3.2.5"
