+++ packages/e2e/e2e-dev-wallet/.env.local
@@ -1,0 +1,1 @@
+WALLETURL=http://localhost:4173
