+++ packages/e2e/e2e-rwa-demo/.env.local
@@ -1,0 +1,1 @@
+WALLETURL=https://wallet.kadena.io
